增加顾客圈选试算接口

+++ yzl-crm-customer-common/src/main/java/cn/net/yzl/crm/customer/mongomodel/member_crowd_group.java
@@ -85,7 +85,7 @@ public class member_crowd_group extends BaseObject {
     private Integer first_order_to_days;
 
     @ApiModelProperty(value = "真正首单金额")
-    private Integer first_order_am;
+    private Double first_order_am;
 
     @ApiModelProperty("最后一次下单时间距离今天多少天")
     private Integer last_order_to_days;
@@ -182,16 +182,16 @@ public class member_crowd_group extends BaseObject {
     private Integer total_amount;
 
     @ApiModelProperty("订单总金额")
-    private Integer order_total_amount;
+    private Double order_total_amount;
 
     @ApiModelProperty("订单应收金额")
-    private Integer order_rec_amount;
+    private Double order_rec_amount;
 
     @ApiModelProperty("订单最高金额")
-    private Integer order_high_am;
+    private Double order_high_am;
 
     @ApiModelProperty("订单最低金额")
-    private Integer order_low_am;
+    private Double order_low_am;
 
     @ApiModelProperty("是否下单: 1是，0否，-1不做条件判断")
     private Integer have_order;
